nav會員icon連結整理完畢

+++ coolbid-app/src/components/member/logout.js
@@ -5,44 +5,11 @@
 /* eslint curly: 2, quotes: ["error", "double"] */
 /* eslint eqeqeq: "off", curly: "error" */
 
-import React from 'react'
-import axios from 'axios'
-import Button from '@material-ui/core/Button'
-import './log.css'
-import styled from '@emotion/styled'
-
-const LoginContainer = styled.div`
-  width: 25%;
-  height: 30rem;
-  margin: 20rem auto;
-  color: grey;
-`
-export default function Logout(props) {
-  axios.defaults.withCredentials = true
-  const handlelogout = () => {
-    axios.get('http://localhost:3001/logout').then((e) => {
-      alert(e.data)
-      console.log(e)
-      // 清sessionstorage
-      window.sessionStorage.clear()
-      window.location.href = 'http://localhost:3000/'
-    })
+export default function Logout() {
+  if (window.sessionStorage.getItem('userinfo')) {
+    // 清sessionstorage
+    window.sessionStorage.clear()
+    alert('登出成功')
+    window.location.href = 'http://localhost:3000/'
   }
-
-  axios.get('http://localhost:3001/member/signin').then((e) => {
-    console.log(e.data.loggedIn)
-  })
-
-  return (
-    <LoginContainer>
-      <Button
-        className="submitBtn"
-        variant="contained"
-        color="primary"
-        onClick={handlelogout}
-      >
-        登出
-      </Button>
-    </LoginContainer>
-  )
 }
